Replace no longer maintaned mysql-python
In anticipation of a move to python3, the mysql-python dependency is
replaced with a newer, maintained, mysqlclient.
As the new package is available inside a OSA repo, holland will
be installed in a separate venv moving forward.
The alternative client, pymysql is not possible due to issue
https://github.com/PyMySQL/PyMySQL/issues/548

Holland is also updated to v1.1.8 with this change.

Closes-Bug: FLEEK-223